+++ packages/picasso/src/ScrollMenu/ScrollMenu.tsx
@@ -11,11 +11,8 @@ import { StandardProps } from '@toptal/picasso-shared'
 import Menu from '../Menu'
 import styles from './styles'
 
-type FocusEventType = (event: React.FocusEvent<HTMLInputElement>) => void
-
 export interface Props extends StandardProps {
   selectedIndex?: number | null
-  onBlur?: FocusEventType
 }
 
 enum Direction {
@@ -37,9 +34,12 @@ const getMoveDirection = (
     : Direction.UP
 }
 
+const preventDefault = (
+  event: React.MouseEvent<HTMLUListElement, MouseEvent>
+) => event.preventDefault()
+
 const ScrollMenu: FunctionComponent<Props> = ({
   selectedIndex,
-  onBlur,
   classes,
   children,
   style
@@ -97,8 +97,8 @@ const ScrollMenu: FunctionComponent<Props> = ({
   }, [firstItemRef, selectedIndex, prevSelectedIndex])
 
   return (
-    <Menu className={classes.menu} style={style}>
-      <div ref={menuRef} className={classes.scrollView} onBlur={onBlur}>
+    <Menu className={classes.menu} style={style} onMouseDown={preventDefault}>
+      <div ref={menuRef} className={classes.scrollView}>
         {renderChildren}
       </div>
     </Menu>